Fix use action in finale carrying over to next map.
As an example, if you pressed the space bar during the finale text at the end of Map06 in Doom II, the text would skip, but the switch at the front of the player at the start of Map07 would be triggered by that. This is fixed. Also, pressing a key will now display the finale text in full, and another press will then advance. Thanks to Jon Krazov & Brad Harding for this.