Correct defect in game restore of targets.
The linked list of thinkers does not get restored in the same order that it was when the game was saved. The P_MobjThinkers gets saved first, followed by the other stuff. Simiarly the P_MobjThinkers get restored first, followed by the rest of the thinkers. So we can only count the P_MobjThinkers when counting the position in the linked list.