Don't use pickup radius if thing moving vertically
The red skull key directly in front of the cyberdemon in MAP15 of Going
Down would get stuck in midair as the platform it is on is lowered. This
is due to code added to P_CheckPosition() in commit https://github.com/jeffdoggett/Doom/commit/27c3c7011a149e3dbe953159bf48093983a923db
Use @https://github.com/bradharding/doomretro idea to use the smaller radius when moving vertically.