Change the bodge for linetype 46.
I originally just made something up so that the shootable switch could be operated multiple times, but the switch texture only changes once. Now that we have generalised doors I can change it.
This seems to be a reasonable compromise between the conficting docs as to wheter it's a G1 or GR.